Revert "Feature/implement locations screen"
