+++ src/components/CharactersPage/CharactersFilters/FiltersInputSearch/FiltersInputSearch.style.ts
