+++ src/components/CharactersPage/CharactersFilters/FiltersInputSearch/index.ts
