--- src/components/LocationsPage/LocationCard/index.ts
@@ -1,1 +1,0 @@
-export * from "./LocationCard";
