--- src/components/LocationsPage/LocationsFilter/index.ts
@@ -1,1 +1,0 @@
-export * from "./LocationsFilter";
