--- src/components/LocationsPage/LocationsList/index.ts
@@ -1,1 +1,0 @@
-export * from "./LocationsList";
