--- src/components/PageTemplate/index.ts
@@ -1,1 +1,0 @@
-export * from "./PageTemplate";
