+++ src/constants/routePath.ts
@@ -1,8 +1,8 @@
 export const routePath = {
   character: "/character",
   characters: "/characters",
-  location: "/location",
-  locations: "/locations",
+  location: "",
+  locations: "",
   episode: "",
   episodes: "",
 };
