+++ src/containers/CharactersPage/CharactersPage.tsx
@@ -1,34 +1,38 @@
 import { useEffect, useState } from "react";
 import { useQuery } from "@apollo/client";
+import { CircularProgress } from "@mui/material";
+
+import { Container } from "@components/Container/Container.style";
 import { CharactersList } from "@components/CharactersPage/CharactersList";
 import { CharactersFilters } from "@components/CharactersPage/CharactersFilters";
 
 import { GET_CHARACTERS } from "@utils/network";
-import {
-  CharacterType,
-  CharactersFilterOptionsType,
-  CharactersSelectOptionsType,
-} from "@/types";
+import { CharacterType, FilterOptionsType, SelectOptionsType } from "@/types";
 
-import logoCharacters from "@/assets/images/logoCharacters.png";
-import { PageTemplate } from "@/components/PageTemplate";
+import {
+  ButtonStyle,
+  ButtonWrapper,
+  CircularContainer,
+  Content,
+  LogoWrapper,
+} from "./CharactersPage.style";
+
+import logoCharacters from "@/assets/images/logo.png";
 
 export const CharactersPage = () => {
   //Хранит уникальные значение силектов
-  const [selectOptions, setSelectOptions] =
-    useState<CharactersSelectOptionsType>({
-      species: [],
-      gender: [],
-      status: [],
-    });
+  const [selectOptions, setSelectOptions] = useState<SelectOptionsType>({
+    species: [],
+    gender: [],
+    status: [],
+  });
   // Фильтры для изменения списка персонажей
-  const [filterOptions, setFilterOptions] =
-    useState<CharactersFilterOptionsType>({
-      name: null,
-      species: null,
-      gender: null,
-      status: null,
-    });
+  const [filterOptions, setFilterOptions] = useState<FilterOptionsType>({
+    name: null,
+    species: null,
+    gender: null,
+    status: null,
+  });
 
   const [items, setItems] = useState([]);
   const [visible, setVisible] = useState(8);
@@ -102,19 +106,32 @@ export const CharactersPage = () => {
   }
 
   return (
-    <PageTemplate
-      FilterComponent={
-        <CharactersFilters
-          selectOption={selectOptions}
-          filterOptions={filterOptions}
-          setFilterOptions={setFilterOptions}
-        />
-      }
-      ListComponent={<CharactersList characters={items} visible={visible} />}
-      logo={logoCharacters}
-      logoAlt="rick and morty"
-      loading={loading}
-      showMoreItems={showMoreItems}
-    />
+    <Container>
+      <div>
+        <LogoWrapper>
+          <img src={logoCharacters} alt="rick and morty" />
+        </LogoWrapper>
+
+        <Content>
+          <CharactersFilters
+            selectOption={selectOptions}
+            filterOptions={filterOptions}
+            setFilterOptions={setFilterOptions}
+          />
+
+          {loading ? (
+            <CircularContainer>
+              <CircularProgress />
+            </CircularContainer>
+          ) : (
+            <CharactersList characters={items} visible={visible} />
+          )}
+
+          <ButtonWrapper>
+            <ButtonStyle onClick={showMoreItems}>LOAD MORE</ButtonStyle>
+          </ButtonWrapper>
+        </Content>
+      </div>
+    </Container>
   );
 };
