--- src/containers/LocationsPage/LocationsPage.style.ts
@@ -1,19 +1,0 @@
-import styled from "styled-components";
-
-export const LogoWrapper = styled.div`
-  margin-top: 26px;
-
-  height: 100%;
-  max-height: 200px;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  img {
-    width: 100%;
-    height: 100%;
-
-    max-width: 326px;
-  }
-`;
