--- src/containers/LocationsPage/index.ts
@@ -1,1 +1,0 @@
-export * from "./LocationsPage";
