--- src/styled_componets/ButtonWrapper.style.ts
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const ButtonWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  margin-top: 60px;
-`;
