--- src/styled_componets/CircularContainer.style.ts
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-
-export const CircularContainer = styled.div`
-  width: 100%;
-  height: 100%;
-  min-height: 400px;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
