--- src/styled_componets/Content.style.ts
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const Content = styled.main`
-  width: 100%;
-`;
