--- src/types/enum.ts
@@ -1,10 +1,0 @@
-export enum LocationType {
-  TYPE = "type",
-  DIMENSION = "dimension",
-}
-
-export enum CharacterType {
-  STATUS = "status",
-  GENDER = "gender",
-  SPECIES = "species",
-}
